\name{ref.ICAR-package}
\alias{ref.ICAR-package}
\alias{ref.ICAR}
\docType{package}

\title{
Objective Bayes Intrinsic Conditional Autoregressive Model for Areal Data
}
\description{
Implements the objective prior for intrinsic conditional autoregressive (ICAR) random effects proposed by Keefe et al. (2018). This model provides an objective Bayesian approach for modeling spatially correlated areal data using an ICAR prior on a vector of spatial random effects.
}

\details{
\tabular{ll}{
Package: \tab ref.ICAR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-10-10\cr
License: \tab GPL (>= 2)\cr
}
}

\references{
\insertRef{keefe2018}{ref.ICAR}

\insertRef{Keefe_2018}{ref.ICAR}
}

\author{
Erica M. Porter, Matthew J. Keefe, Christopher T. Franck, and Marco A.R. Ferreira \cr
Maintainer: Erica M. Porter <ericamp@vt.edu>
}

\examples{
## Refer to the vignette attached to the package,
## and to selected help files.
}
