% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.plot.R
\name{ref.plot}
\alias{ref.plot}
\title{Trace Plots for Parameters in ICAR Model}
\usage{
ref.plot(MCMCchain, X, burnin, num.reg)
}
\arguments{
\item{MCMCchain}{Matrix of MCMC chains for the model parameters.}

\item{X}{Matrix of covariates.}

\item{burnin}{Number of MCMC iterations from \code{MCMCchain} discarded as burn-in.}

\item{num.reg}{Number of regions in the areal data set.}
}
\value{
Trace plots for the fixed effect regression coefficients, the precision parameter,
and the unstructured random effects variance.
}
\description{
This function creates trace plots for the parameters in
the ICAR reference prior model \insertCite{keefe2018}{ref.ICAR}.
}
\examples{
## Refer to the vignette attached to the package.
}
\author{
Erica M. Porter, Matthew J. Keefe, Christopher T. Franck, and Marco A.R. Ferreira
}
